refactor(editor): widen SelectionManager::EntityId from uint32_t to uint64_t
The audit at .claude/knowledge/stub-and-abandoned-features-2026-04-10.md
flagged that HierarchyPanel, InspectorPanel, and SceneView still maintain
their own per-panel selection state instead of using the
SelectionManager singleton. The blocker for the migration was that
SelectionManager::EntityId was uint32_t while the editor's ObjectID
(SceneFileTypes.h:34) is uint64_t — every panel migration would need a
narrowing conversion.

This commit takes the smallest possible step to unblock the future
panel migration: change the EntityId alias to uint64_t. The alias is
the only type that needed updating; std::hash<uint64_t> is a drop-in
for std::hash<uint32_t>, all comparisons widen, and the literal casts
in Tests/TestSelectionManager.cpp (static_cast<EntityId>(20) etc.) work
for any unsigned width.

m_callbackId / m_nextCallbackId / RemoveCallback() stay uint32_t —
they are unrelated callback handles, not entity IDs.

The full panel migration (replacing HierarchyPanel::m_selectedObjects
etc. with SelectionManager::GetInstance() calls) is still pending; it
is a multi-file refactor estimated at ~80-100 call site updates and
needs its own branch. This commit only unblocks it.

+++ SparkEditor/Source/Panels/SelectionManager.h
@@ -61,8 +61,19 @@
 namespace SparkEditor
 {
 
-    /** @brief Entity identifier for selection purposes */
-    using EntityId = uint32_t;
+    /**
+     * @brief Entity identifier for selection purposes.
+     *
+     * Matches `ObjectID` from `SparkEditor/Source/SceneSystem/SceneFileTypes.h`
+     * (`using ObjectID = uint64_t`) so that editor panels like
+     * `HierarchyPanel` and `InspectorPanel`, which store `ObjectID`
+     * selection state, can migrate to this singleton without a narrowing
+     * conversion. The full panel-level migration (replacing per-panel
+     * `m_selectedObjects` etc. with `SelectionManager::GetInstance()`) is
+     * a separate follow-up — see
+     * `.claude/knowledge/engine-next-steps-2026-04-10.md` Phase B item 2.
+     */
+    using EntityId = uint64_t;
 
     /** @brief Invalid entity sentinel */
     constexpr EntityId INVALID_ENTITY = 0;
